+++ tests/testthat/test-09-report.R
@@ -35,7 +35,7 @@ test_that("errors", {
 })
 
 test_that("rendering error", {
-  # should give a warning and returnt he path to the saved qmd
+  # should give a warning and return the path to the saved qmd
   paper <- read(demoxml())
   modules <- c("bad-report")
   output_format <- "html"
@@ -60,6 +60,43 @@ test_that("report return list", {
   expect_equal(names(paper_report), modules)
   expect_s3_class(paper_report[[1]], "metacheck_module_output")
   expect_equal(paper_report$stat_p_exact$module, "stat_p_exact")
+})
+
+test_that("report paperlist", {
+  paper <- psychsci[1:2]
+  modules <- c("stat_p_exact", "marginal")
+  output_file <- withr::local_tempfile(pattern = "_", fileext = ".qmd")
+  output_format <- "qmd"
+  paper_report <- report(paper, modules, output_file, output_format)
+  expect_equal(length(paper_report), 2)
+  expect_equal(names(paper_report), names(paper))
+  expect_equal(names(paper_report[[1]]), modules)
+  expect_equal(names(paper_report[[2]]), modules)
+  sp1 <- attr(paper_report[[1]], "save_path")
+  expect_true(file.exists(sp1))
+  sp2 <- attr(paper_report[[2]], "save_path")
+  expect_true(file.exists(sp2))
+  sp <- attr(paper_report, "save_path")
+  expect_equal(names(sp), names(paper))
+  expect_equal(sp[[1]], sp1)
+  expect_equal(sp[[2]], sp2)
+
+  output_file1 <- withr::local_tempfile(fileext = ".qmd")
+  output_file2 <- withr::local_tempfile(fileext = ".qmd")
+  pr1 <- report(paper[1], modules, output_file1, output_format)
+  pr2 <- report(paper[[2]], modules, output_file2, output_format)
+  expect_true(file.exists(output_file1))
+  expect_true(file.exists(output_file2))
+
+  expect_equal(paper_report[[1]], pr1, ignore_attr = TRUE)
+  expect_equal(paper_report[[2]], pr2, ignore_attr = TRUE)
+
+  # vector of output_files
+  output_file <- c(withr::local_tempfile(fileext = ".qmd"),
+                   withr::local_tempfile(fileext = ".qmd"))
+  paper_report <- report(paper, modules, output_file, output_format)
+  expect_true(file.exists(output_file[[1]]))
+  expect_true(file.exists(output_file[[2]]))
 })
 
 test_that("render qmd", {
@@ -232,5 +269,29 @@ test_that("module_report howitworks", {
 })
 
 
+test_that("report_module_run", {
+  expect_true(is.function(metacheck::report_module_run))
+  expect_no_error(helplist <- help(report_module_run, metacheck))
 
+  paper <- read(demoxml())
+  modules <- "all_p_values"
+  mo <- report_module_run(paper, modules)
+  expect_equal(names(mo), modules)
 
+  modules <- c("stat_p_nonsig", "stat_p_exact")
+  mo <- report_module_run(paper, modules)
+  expect_setequal(names(mo), modules)
+})
+
+test_that("report_qmd", {
+  expect_true(is.function(metacheck::report_qmd))
+  expect_no_error(helplist <- help(report_qmd, metacheck))
+
+  paper <- read(demoxml())
+  modules <- "stat_p_nonsig"
+  mo <- report_module_run(paper, modules)
+  report_text <- report_qmd(mo, paper)
+  expect_true(grepl("MetaCheck Report", report_text))
+  expect_true(grepl(paper$info$title, report_text))
+  expect_true(grepl("Non-Significant P Value Check", report_text, fixed = TRUE))
+})