+++ NEWS.md
@@ -1,3 +1,10 @@
+# metacheck 0.0.0.9067
+
+* `reports()` now takes a paperlist and makes a report from each
+* New `report_module_run()` and `report_qmd()` break down the `report()` function to allow separation of module output lists and creation of QMD report from them (might be changed to internal functions).
+* Ability to select returned columns in `crossref_query()`
+* Module "ref_accuracy" now returns info for references with missing DOIs that were found by ref_doi_check
+
 # metacheck 0.0.0.9066
 
 * `lmm()` allows you to set the model to any provider or provider/model supported by ellmer (must have appropriate *****_API_KEY set in your Renviron)